+++ spec/guard/resque_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe Guard::Delayed do
-  subject { Guard::Delayed.new }
+describe Guard::Resque do
+  subject { Guard::Resque.new }
 
   describe "start" do
     before do
-      subject = Guard::Delayed.new
+      subject = Guard::Resque.new
     end
 
     after do
@@ -14,7 +14,7 @@ describe Guard::Delayed do
 
     it "should create a pid in /tmp/pids" do
       subject.start
-      Dir.glob("tmp/pids/delayed_job*.pid").should_not be_empty
+      Dir.glob("tmp/pids/resque*.pid").should_not be_empty
     end
   end
 end